Add generator script
