+++ Project.toml
@@ -1,16 +1,14 @@
 name = "MiniFB"
 uuid = "cc649173-1898-4bb8-8a16-ca5f88bd87d9"
 authors = ["Avik Sengupta"]
-version = "0.1.1"
+version = "0.2.0"
 
 [deps]
 Colors = "5ae59095-9a9b-59fe-a467-6f913c188581"
-CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
 MiniFB_jll = "b93afbca-07d8-5443-9979-bb4221978155"
 
 [compat]
 Colors = "0.10, 0.11, 0.12"
-CEnum = "0.4"
 julia = "1.3"
 
 [extras]
