+++ example/Noise.jl
@@ -1,6 +1,8 @@
 
 
 using MiniFB
+using MiniFB.LibMiniFB
+
 # A simple function that renders some noise to a window. Demonstrates how to create a window and render a buffer into it using MiniFB
 function noise()
     WIDTH = 800
@@ -18,7 +20,7 @@ function noise()
             seed = seed >> 1;
             seed = seed | (carry << 30);
             noise = noise & 0xFF;
-            g_buffer[i] = mfb_rgb(noise, noise, noise); 
+            g_buffer[i] = mfb_rgb(noise, noise, noise);
         end
         state = mfb_update(window, g_buffer);
         if state != MiniFB.STATE_OK
