+++ gen/.gitignore
@@ -1,0 +1,4 @@
+MiniFB.h
+MiniFB_enums.h
+MiniFB_cpp.h
+MiniFB_ios.h
