+++ gen/Project.toml
@@ -1,0 +1,7 @@
+[deps]
+Clang = "40e3b903-d033-50b4-a0cc-940c62c95e31"
+Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
+MiniFB_jll = "b93afbca-07d8-5443-9979-bb4221978155"
+
+[compat]
+Clang = "0.14"
