+++ src/MiniFB.jl
@@ -2,14 +2,12 @@ module MiniFB
 using Colors
 using MiniFB_jll
 
-using CEnum
+include("LibMiniFB.jl")
+using .LibMiniFB
 
-include("ctypes.jl")
-export Ctm, Ctime_t, Cclock_t
-
-include("libminifb_common.jl")
-include("libminifb_api.jl")
 include("libminifb_docs.jl")
 include("helpers.jl")
+
+export mfb_rgb
 
 end
