--- src/ctypes.jl
@@ -1,7 +1,0 @@
-## TODO: pending https://github.com/JuliaLang/julia/issues/29420
-# this one is suggested in the issue, but it looks like time_t and tm are two different things?
-# const Ctime_t = Base.Libc.TmStruct
-
-const Ctm = Base.Libc.TmStruct
-const Ctime_t = UInt
-const Cclock_t = UInt
